Implemented CLI for wrapper
