+++ neat/bacterial_wrapper/__init__.py
@@ -1,0 +1,4 @@
+"""
+Load modules needed for other parts of the program
+"""
+from .runner import *
